refactor: etcd != remote.
