+++ config/remote/etcd/etcd.go
@@ -1,4 +1,5 @@
-package remote
+// Package etcd allows the core package to bootstrap its configuration from an etcd server.
+package etcd
 
 import (
 	"context"
@@ -8,23 +9,23 @@ import (
 	"go.etcd.io/etcd/client/v3"
 )
 
-// Remote is a core.ConfProvider and contract.ConfigWatcher implementation to read and watch remote config key.
+// ETCD is a core.ConfProvider and contract.ConfigWatcher implementation to read and watch remote config key.
 // The remote client uses etcd.
-type Remote struct {
-	key         string
+type ETCD struct {
+	key          string
 	clientConfig *clientv3.Config
 }
 
-// Provider create a *Remote
-func Provider(key string, clientConfig *clientv3.Config) *Remote {
-	return &Remote{
-		key:         key,
+// Provider create a *ETCD
+func Provider(key string, clientConfig *clientv3.Config) *ETCD {
+	return &ETCD{
+		key:          key,
 		clientConfig: clientConfig,
 	}
 }
 
 // ReadBytes reads the contents of a key from etcd and returns the bytes.
-func (r *Remote) ReadBytes() ([]byte, error) {
+func (r *ETCD) ReadBytes() ([]byte, error) {
 	client, err := clientv3.New(*r.clientConfig)
 	if err != nil {
 		return nil, err
@@ -43,7 +44,7 @@ func (r *Remote) ReadBytes() ([]byte, error) {
 }
 
 // Read is not supported by the remote provider.
-func (r *Remote) Read() (map[string]interface{}, error) {
+func (r *ETCD) Read() (map[string]interface{}, error) {
 	return nil, errors.New("remote provider does not support this method")
 }
 
@@ -51,7 +52,7 @@ func (r *Remote) Read() (map[string]interface{}, error) {
 // will be called. note the reload function should not just load the changes made within this key, but rather
 // it should reload the whole config stack. For example, if the flag or env takes precedence over the config
 // key, they should remain to be so after the key changes.
-func (r *Remote) Watch(ctx context.Context, reload func() error) error {
+func (r *ETCD) Watch(ctx context.Context, reload func() error) error {
 	client, err := clientv3.New(*r.clientConfig)
 	if err != nil {
 		return err
@@ -74,4 +75,3 @@ func (r *Remote) Watch(ctx context.Context, reload func() error) error {
 		}
 	}
 }
-
